hotfix: config file database + gitignore
