--- alembic/versions/0b7f349e19fc_initial_schema.py
@@ -1,32 +1,0 @@
-"""Initial schema
-
-Revision ID: 0b7f349e19fc
-Revises: 4eed3147b5f3
-Create Date: 2025-06-04 11:19:12.304759
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '0b7f349e19fc'
-down_revision: Union[str, None] = '4eed3147b5f3'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
