--- alembic/versions/3b64a2fca9b3_initial_schema.py
@@ -1,139 +1,0 @@
-"""Initial schema
-
-Revision ID: 3b64a2fca9b3
-Revises: 53830c99888a
-Create Date: 2025-06-17 20:51:44.032782
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '3b64a2fca9b3'
-down_revision: Union[str, None] = '53830c99888a'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('salles',
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('nom', sa.String(), nullable=False),
-    sa.Column('capacite', sa.Integer(), nullable=True),
-    sa.Column('localisation', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('reservations',
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('salle_id', sa.String(), nullable=False),
-    sa.Column('utilisateur', sa.String(), nullable=False),
-    sa.Column('date', sa.Date(), nullable=False),
-    sa.Column('heure', sa.Time(), nullable=False),
-    sa.ForeignKeyConstraint(['salle_id'], ['salles.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.drop_table('commandes')
-    op.drop_table('implantations')
-    op.drop_table('receptions')
-    op.drop_table('articles')
-    op.drop_table('agents')
-    op.drop_table('emplacements')
-    op.drop_table('lignes_commandes')
-    op.drop_table('missions')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('missions',
-    sa.Column('id', sa.VARCHAR(), nullable=False),
-    sa.Column('type', sa.VARCHAR(length=11), nullable=False),
-    sa.Column('etat', sa.VARCHAR(length=8), nullable=False),
-    sa.Column('article_id', sa.VARCHAR(), nullable=False),
-    sa.Column('source_id', sa.VARCHAR(), nullable=True),
-    sa.Column('destination_id', sa.VARCHAR(), nullable=True),
-    sa.Column('quantite', sa.INTEGER(), nullable=False),
-    sa.Column('agent_id', sa.VARCHAR(), nullable=True),
-    sa.Column('date_creation', sa.DATETIME(), nullable=True),
-    sa.Column('date_execution', sa.DATETIME(), nullable=True),
-    sa.Column('urgent', sa.BOOLEAN(), nullable=False),
-    sa.ForeignKeyConstraint(['agent_id'], ['agents.id'], ),
-    sa.ForeignKeyConstraint(['article_id'], ['articles.id'], ),
-    sa.ForeignKeyConstraint(['destination_id'], ['emplacements.id'], ),
-    sa.ForeignKeyConstraint(['source_id'], ['emplacements.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('lignes_commandes',
-    sa.Column('id', sa.VARCHAR(), nullable=False),
-    sa.Column('commande_id', sa.VARCHAR(), nullable=True),
-    sa.Column('article_id', sa.VARCHAR(), nullable=True),
-    sa.Column('quantite', sa.INTEGER(), nullable=False),
-    sa.ForeignKeyConstraint(['article_id'], ['articles.id'], ),
-    sa.ForeignKeyConstraint(['commande_id'], ['commandes.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('emplacements',
-    sa.Column('id', sa.VARCHAR(), nullable=False),
-    sa.Column('code', sa.VARCHAR(), nullable=False),
-    sa.Column('type', sa.VARCHAR(length=11), nullable=False),
-    sa.Column('capacite_poids_kg', sa.FLOAT(), nullable=False),
-    sa.Column('capacite_volume_m3', sa.FLOAT(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('code')
-    )
-    op.create_table('agents',
-    sa.Column('id', sa.VARCHAR(), nullable=False),
-    sa.Column('nom', sa.VARCHAR(), nullable=False),
-    sa.Column('email', sa.VARCHAR(), nullable=False),
-    sa.Column('actif', sa.BOOLEAN(), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email')
-    )
-    op.create_table('articles',
-    sa.Column('id', sa.VARCHAR(), nullable=False),
-    sa.Column('sku', sa.VARCHAR(), nullable=False),
-    sa.Column('designation', sa.VARCHAR(), nullable=False),
-    sa.Column('categorie', sa.VARCHAR(length=11), nullable=False),
-    sa.Column('poids_kg', sa.FLOAT(), nullable=False),
-    sa.Column('volume_m3', sa.FLOAT(), nullable=False),
-    sa.Column('date_peremption', sa.DATE(), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('sku')
-    )
-    op.create_table('receptions',
-    sa.Column('id', sa.VARCHAR(), nullable=False),
-    sa.Column('article_id', sa.VARCHAR(), nullable=False),
-    sa.Column('quantite', sa.INTEGER(), nullable=False),
-    sa.Column('fournisseur', sa.VARCHAR(), nullable=False),
-    sa.Column('date_reception', sa.DATETIME(), nullable=True),
-    sa.Column('emplacement_id', sa.VARCHAR(), nullable=False),
-    sa.ForeignKeyConstraint(['article_id'], ['articles.id'], ),
-    sa.ForeignKeyConstraint(['emplacement_id'], ['emplacements.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('implantations',
-    sa.Column('id', sa.VARCHAR(), nullable=False),
-    sa.Column('article_id', sa.VARCHAR(), nullable=False),
-    sa.Column('emplacement_id', sa.VARCHAR(), nullable=False),
-    sa.Column('quantite', sa.INTEGER(), nullable=False),
-    sa.Column('seuil_minimum', sa.INTEGER(), nullable=False),
-    sa.ForeignKeyConstraint(['article_id'], ['articles.id'], ),
-    sa.ForeignKeyConstraint(['emplacement_id'], ['emplacements.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('commandes',
-    sa.Column('id', sa.VARCHAR(), nullable=False),
-    sa.Column('reference', sa.VARCHAR(), nullable=False),
-    sa.Column('etat', sa.VARCHAR(length=9), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('reference')
-    )
-    op.drop_table('reservations')
-    op.drop_table('salles')
-    # ### end Alembic commands ###
