--- alembic/versions/4eed3147b5f3_initial_schema.py
@@ -1,32 +1,0 @@
-"""Initial schema
-
-Revision ID: 4eed3147b5f3
-Revises: e6c007a09a72
-Create Date: 2025-06-04 11:13:37.414068
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '4eed3147b5f3'
-down_revision: Union[str, None] = 'e6c007a09a72'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
