--- alembic/versions/53830c99888a_initial_schema.py
@@ -1,32 +1,0 @@
-"""Initial schema
-
-Revision ID: 53830c99888a
-Revises: 0b7f349e19fc
-Create Date: 2025-06-06 13:55:15.031093
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '53830c99888a'
-down_revision: Union[str, None] = '0b7f349e19fc'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('missions', sa.Column('urgent', sa.Boolean(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('missions', 'urgent')
-    # ### end Alembic commands ###
