+++ alembic/versions/75841b955e98_initial_schema.py
@@ -1,0 +1,51 @@
+"""Initial schema
+
+Revision ID: 75841b955e98
+Revises: 
+Create Date: 2025-06-19 15:37:19.160995
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '75841b955e98'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('salles',
+    sa.Column('id', sa.String(), nullable=False),
+    sa.Column('nom', sa.String(), nullable=False),
+    sa.Column('capacite', sa.Integer(), nullable=True),
+    sa.Column('localisation', sa.String(), nullable=True),
+    sa.Column('disponible', sa.Boolean(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('nom', name='uq_salle_nom')
+    )
+    op.create_table('reservations',
+    sa.Column('id', sa.String(), nullable=False),
+    sa.Column('salle_id', sa.String(), nullable=False),
+    sa.Column('utilisateur', sa.String(), nullable=False),
+    sa.Column('date', sa.Date(), nullable=False),
+    sa.Column('heure', sa.Time(), nullable=False),
+    sa.Column('commentaire', sa.String(), nullable=True),
+    sa.ForeignKeyConstraint(['salle_id'], ['salles.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('reservations')
+    op.drop_table('salles')
+    # ### end Alembic commands ###
